+++ fancontrol/__init__.py
@@ -1,4 +1,4 @@
 """A fan controller for Linux, in the shape of FanControl for Windows."""
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 __all__ = ["__version__"]